+++ octue/resources/pubsub/__init__.py
@@ -1,0 +1,5 @@
+from .subscription import Subscription
+from .topic import Topic
+
+
+__all__ = ["Subscription", "Topic"]
